REL-2886 Move Browser and Platform class to the shared gui component #time 1h
