+++ bundle/edu.gemini.pit/src/main/scala/edu/gemini/pit/ui/action/BrowseAction.scala
@@ -2,9 +2,10 @@ package edu.gemini.pit.ui.action
 
 import java.net.URI
 import javax.swing.Action.ACCELERATOR_KEY
-import javax.swing.{KeyStroke, AbstractAction}
+import javax.swing.{AbstractAction, KeyStroke}
 import java.awt.event.ActionEvent
-import edu.gemini.pit.ui.util.Browser
+
+import edu.gemini.shared.gui.Browser
 
 class BrowseAction(uri: URI, text: String, vkey:Int = 0, modifiers: Int = 0) extends AbstractAction(text) {
   if (vkey != 0) putValue(ACCELERATOR_KEY, KeyStroke.getKeyStroke(vkey, modifiers))
