+++ bundle/edu.gemini.pit/src/main/scala/edu/gemini/pit/ui/util/ShellAction.scala
@@ -1,8 +1,9 @@
 package edu.gemini.pit.ui.util
 
 import java.awt.event.ActionEvent
-import javax.swing.{ KeyStroke, Action, AbstractAction }
+import javax.swing.{AbstractAction, Action, KeyStroke}
 
+import edu.gemini.shared.Platform
 import edu.gemini.ui.workspace.scala._
 
 abstract class ShellAction[A](shell: RichShell[A], caption: String, key: Option[Int] = None, mask: Int = 0) extends AbstractAction(caption) with (() => Unit) {
